Correcion de errores

+++ MyBatis_Introduction_VideoRental/src/main/java/edu/eci/cvds/sampleprj/dao/ClienteDAO.java
@@ -8,7 +8,7 @@ public interface ClienteDAO {
 
    public void save(Cliente cliente) throws PersistenceException;
 
-   public Cliente load(int id) throws PersistenceException;
+   public Cliente load(long id) throws PersistenceException;
 
    public List<Cliente> consultarClientes() throws PersistenceException;
